requirements.txt
fastapi
google-auth
httpx
uvicorn
